GR_Refactor-Mixer-&-Gameplay-Config
Se saco la responsabilidad del mixer y gamplay config del gamemanager.
Se dejo la responsabilidad de poner el screendeath al gamemanager suscribiendose al ondead del damageable del player
