+++ Assets/_Scripts/Managers/GameplayConfigManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 45d98254e66bf724ebb95c653d7254fc
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
