+++ Assets/_Scripts/Managers/MixerManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 47919a1f94e57d041b380655ffc1ef26
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
